Fix a bug in the `prepare` command with missing Git commits
If you run `git show` for a commit that isn't present (e.g. a merge
commit created on GitHub), you'd get an error from the `prepare` command:

    fatal: ambiguous argument '666929a': unknown revision or path not in the working tree.
    Use '--' to separate paths from revisions, like this:
    'git <command> [<revision>...] -- [<file>...]'

This patch does a `git fetch` (not a `pull`) to get all the commits on
the remote, and also handles a missing commit more gracefully.

Also tweak the formatting so you actually get the subject line, and not
the entire commit body (oops).